refactor(actions): useMarkdownActions hook + pendingHandledRef
- Логика applyMarkdownAction и placeholder'ов вынесена из App.tsx
  в новый hook src/hooks/useMarkdownActions.ts. Также появился
  insertAssetAction, инкапсулирующий цепочку pickAndFormatAsset →
  applyMarkdownAction.
- Toolbar теперь вызывает insertAssetAction() напрямую вместо
  диспатча синтетического Ctrl+Shift+A через KeyboardEvent. Кнопка
  больше не сломается при изменении хоткея и тестируется тривиально.
- App.tsx: pendingHandledRef = useRef(false) — явная StrictMode-
  защита для open-via-association эффекта. Логика больше не
  полагается на побочный эффект Mutex.take() на Rust-стороне.
