+++ src/components/Toolbar/Toolbar.tsx
@@ -2,6 +2,7 @@ import { useState, useEffect } from 'react';
 import { useFile } from '../../hooks/useFile';
 import { useSettings } from '../../hooks/useSettings';
 import { useTheme } from '../../hooks/useTheme';
+import { useMarkdownActions } from '../../hooks/useMarkdownActions';
 import { useAppStore } from '../../stores/appStore';
 import { getTranslations } from '../../i18n';
 import { HelpDialog } from '../Help/HelpDialog';
@@ -21,6 +22,7 @@ export function Toolbar() {
   const [showHelp, setShowHelp] = useState(false);
   const { open, save, saveAs, reload, filePath } = useFile();
   const { fontFamily, fontSize, language, pageWidth, changeFontFamily, changeFontSize, changeLanguage, changePageWidth } = useSettings();
+  const { insertAssetAction } = useMarkdownActions();
   const [pageWidthDraft, setPageWidthDraft] = useState(String(pageWidth));
 
   // Синхронизация при внешнем изменении (загрузка настроек)
@@ -77,7 +79,7 @@ export function Toolbar() {
           </button>
           <button
             className="toolbar-btn"
-            onClick={() => window.dispatchEvent(new KeyboardEvent('keydown', { key: 'A', code: 'KeyA', ctrlKey: true, shiftKey: true, bubbles: true }))}
+            onClick={() => insertAssetAction()}
             disabled={!filePath}
             title={t.insertAssetTooltip}
           >
